external-arago-toolchain: bring back SUPPORTED file, required by eglibc
The initial testing was done with the SUPPORTED file, but then it was removed,
as it's not directly used by the external toolchain recipe. But it is being used
by eglibc packaging routines inside the framework and breaks w/o it. Since the
file was previously placed in the work directory, it didn't break until the build
folder was completely removed for a clean build or anyone starting from scratch.

Signed-off-by: Denys Dmytriyenko <denis@denix.org>