arago.conf: pin down udev version from oe-core, instead of one from meta-oe
meta-oe brings in newer, systemd-enhanced version of udev-182. In "denzil"
branch it seems to completely depend on systemd, while in master the hard
dependency is removed, but sysvinit support is still somewhat broken, as
initscript for udev appears to be missing. It should be possible to fix it
in master, but for "denzil" it would need the meta-systemd split anyway, so
the easiest solution is to use the older udev-164 from oe-core, that works
fine with sysvinit.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>