python-m2crypto: fix wrong checksums and failing patches
When porting this recipe from classic-oe, the patches lost all their extra
whitespaces and tabs, making them to not apply anymore. Copy the original
patches directly from classic-oe.

Also, the checksums were changed for some reason during porting, revert them
back to the same as in classic-oe, which are the correct ones.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>