arago-prefs.inc: add selection between wilink versions
Original by Eyal Reizer.

Add a configuration variable for selecting between wilink6 and
wilink8 support.

Based on this variable the right version will be picked for the
recipes which are different between wilink versions.

The default is currently set to "WILINK-6" until the next SDK release
which is going to switch to using WILINK-8.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>