tisdk-install: Move SDK installation script
Several steps are required before an SDK image is ready to be used.
1. Extract toolchain sdk (linuxdevkit.sh)
2. Update Rules.make variables.
3. Update CCS project information

The script that sdk-install.sh is based on was previously packaged and ran by
the SDK installed (based on InstallJammer).

By moving the logic outside of the installer. OE users can now create
a fully functioning SDK without any outside dependencies by simply running
sdk-install located in the root of the SDK image.

Signed-off-by: Franklin S. Cooper Jr <fcooper@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>