tisdk-image: Fix manifest generation for nativesdk.
The software manifest is created by parsing the *.control files in the images
/var/lib/opkg directory. For the nativesdk sysroot, there are some cases where
the control files are not present. The reason for this and the specific
circumstances of when this occurs has not yet been root caused.

When this does happen, the "Development Host Content" tables of the manifest
are not valid. However, the package indexes are present. These contain all of
the information in the control files, only concatenated together in a large
file. This patch will split up these package indexes into individual control
files which will allow the previous parsing to produce a valid output.

Here is an example of the invalid output when the control files are not present:

<h2><u>Development Host Content</u></h2>
<p>This table describes any software being delivered that is expected to run on
a Development Host, instead of the target device.  Some of this software may be
licensed under GPLv3 but it is not expected to be shipped as a product.</p>

<table border=1 cellspacing=1 cellpadding=1 width=80%>
<tr bgcolor=#c0c0c0  color=white>
    <td><b>Software Name</b></td>
    <td><b>Version</b></td>
    <td><b>License</b></td>
    <td><b>Location</b></td>
    <td><b>Delivered As</b></td>
    <td><b>Modified by TI</b></td>
    <td><b>Obtained from</b></td>
</tr>

<tr>
    <td> </td>
    <td></td>
    <td ></td>
    <td>__.ipk</td>
    <td>Binary</td>
    <td>No</td>
    <td></td>
</tr>

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>