weston: Add custom profile into targetfs
There are different set of requirements on the bootup behavior
across different platforms. To name a few:
 * J6, the required behavior is console/command prompt
 * AM57xx, it will be the matrix-gui (QT5 on Wayland)

weston-init recipe addresses the needs of AM57xx.
This patch is needed for J6/J6-Eco etc.

Note that if there are customers on AM57xx who need the default behavior
to be console/command prompt, this patch will address those needs.
It does not conflict with weston-init.

Signed-off-by: Karthik Ramanan <a0393906@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>