meta-arago-toolchain: fix relocation of default library paths
Pass the actual SDKPATH to relocation script instead of what we want the
default to be (confusingly named SDKPATH_REAL). Since most of the binaries
use relative library paths (via $ORIGIN) anyways, it wasn't causing much
trouble. But git uses a set of commands executed from /usr/lib/git/git-core
and those binaries fail to locate libraries relative to that path and fall
back to default library paths, but those were not properly updated by the
relocation script, as it was looking for SDKPATH_REAL instead of SDKPATH.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>