dual-camera-demo: Re-enable on AM4 to run outside of Weston
The dual camera demo stopped working when Wayland/DRM-based graphics support
was introduced on AM4 and therefore this demo was disabled. This demo does not
work with Wayland/DRM-based graphics system because it invokes DRM APIs to
demonstrate DRM mode setings and overlay operations and those operations
are denied because there should be only one DRM master.

The simple solution is to update the demo script to run this QT5-based demo
program over linuxfb platform and therefore there is not another DRM master
running.

Signed-off-by: Eric Ruei <e-ruei1@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>