sourceipk: fix git preservation against bitbake 1.30
With previous versions of bitbake, the git repo in the workdir would
point its remote/origin to the local clone in the DL_DIR. The
sourceipk class would thus need to go to the local clone to get the
true upstream remote.

With bitbake commit 80ecd1c54d4c748cee3a7ce0d64013a346e7671e
(branch: 1.30), the workdir repo now points to the true upstream repo

This fixes the sourceipk git preservation that was broken due to this
change in bitbake. Additional changes were required in the limit git
history feature to ensure that the preserved git points to the commit
used within the build.

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>