external-linaro-toolchain: reorder PACKAGES so libgcc-dev picks libgcc.a
Recent update to the upstream external-linaro-toolchain.bb in meta-linaro
changed the order of PACKAGES and libgcc-dev is now towards the end of the
list and has no chance of picking libgcc.a. Since native compiler on the
target uses libgcc.a and expects it to come from libgcc-dev package, the
fix is to move it up the PACKAGES list.

Moreover, with gcc6, linaro prebuilt toolchain stopped packaging libgcc.a
inside the target sysroot, so pick it up from gcc lib dir instead.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>