wayland-ivi-extension: add recipe
This recipe is primarily adapted from the meta-agl.
URL reference: https://goo.gl/m0LLmT

The recipe brings in the capability to use the ivi-shell
feature of the Weston and provides utilities to manage
the screens, layers and surfaces.

force-type-conversion.patch overlayed in the recipe is required
to address a compilation issue seen only with TI graphics
stack.

The other patches are required on top of the 1.11.0 version to
fully use the multiple screen features on TI platforms.

Signed-off-by: Karthik Ramanan <a0393906@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>