netboot-initramfs: Make it possible to use NFS as the rootfs with non-builtin drivers
USB drivers are usually built as modules. This prevents using NFS as the
rootfs over a USB because the in-kernel support is not yet available at
the end of the kernel init.

To circumvent the problem, a minimal initramfs can be used to probe the
missing modules and finally setup the NFS mount.

The initramfs contains the required modules, udev and a
init script.
The script's actions are (in execution order):
* starts udev and trigger udev's enumeration
* parses the kernel command line to extract the NFS kernel options
* configure the ethernet device
* mount the NFS
* execute switch_root

The list of kernel module is defined by the variable NETBOOT_KERNEL_MODULES.
It contains the list of modules separated by space without the ".ko"
extension.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>