netboot-initd: Fix parsing error in init script
The init script parses the kernel command line to extract information to
mount the NFS root filesystem. If a parameter starts with 'root', 'ip' or
'nfsroot' but is not one of those parameters (for example rootwait starts
with 'root' but is not the 'root' parameter), then the parsing fails.
Fix this by also looking for the '=' sign after the parameter name.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>