chromium-wayland: adjust build flags to disable debug and fix size issue
Normally binaries are built with debug symbols and then those are stripped out
into corresponding -dbg packages for debugging purposes.

Chromium builds in a lot of components and that leads to a large binary size:
* with gcc7 it was 3.9GB with debugging info, stripped down to 80MB w/o it
* with gcc8 it became 4.3GB and that is larger than 32-bit addressing, so it
becomes corrupted and cannot be stripped or executed

By disabling generation of the debugging info in the first place, it is now
possible to get a working chromium, but w/o the ability to debug it.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>