sourceipk: remove custom package creation
Instead of manually creating the source "ipk", use the built in
pacakge_write_* tasks. This is done by constructing the sourceipk
package contents in a staging area. Then, as a PACKAGESPLITFUNCS,
copy the contents of the staging area into the appropriate
packages-split directory so that it can get used during the package
write task. Disable all QA checks on this package.

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>