kernel-fitimage: Use compressed ramdisks in FIT images if available
This is an adaptation of the following commit from upstream OE-Core:
https://git.openembedded.org/openembedded-core/commit/?id=842ad404b36e00c89f615a3f7db4a2d30062effa

Previously, it was calling copy_initramfs from kernel.bbclass, which would
decompress the ramdisk. To make sure we use as little space as possible, we
used to re-compress it into XZ. Now it just uses the original format of the
ramdisk, either uncompressed or in any of the supported compression methods.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>