arago-prefs: Set virtual/libgbm as provided by ti-sgx-ddk-um
The GBM library is no longer a separate package, like with Mesa and
RGX driver it is provided by the GPU driver. Cleanup the setting of
this to make it more clear.

Also drop virtual/libgbm-dev, dev packages are not valid when used like
this or is this needed by anyone to begin with.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>