gstreamer1.0*-bad: Update Parsers Patch for k3
Remove the part of the parsers patch that changes the
width and height during create for h264 streams. This
change was needed only for gst-ducati plugin and gst-ducati
plugin is not used for k3. Additionally the change was
causing artifacts for some videos when used with v4l2h264dec
plugin.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>