Rules.make: Initialie DESTDIR and EXEC_DIR conditionally
DESTDIR and EXEC_DIR in the Rules.make point to the paths where
the files should be installed.

Initialize the conditionally so that command line override is
allowed. This is useful when you want to switch between NFS rootfs
and SD card rootfs.
e.g. Following can be done now from installer.

$> DESTDIR=/path/to/sd/card make install #custom installation
$>                          make install #NFS installation

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>