docker: install init script if sysvinit is enabled
The docker recipe in meta-virtualization has conflicting logic for
when to install and when to enable the docker initscript. When both
systemd and sysvinit are in DISTRO_FEATURES, this causes the docker
postinst to fail, which now fails the rootfs task.

Here we install the initscript only based on if sysvinit is in the
DISTRO_FEATURES.

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>