opencl: major refactor
For opencl, use the runtime package to pull in the minimal set of
dependencies needed to run opencl applications. The applications must
use precompiled kernels. Use the main package to pull in the full set
of tools, including clocl and the accelerator toolchains.

For the opencl-examples, the offline-compile package are the examples
which provide precompiled kernels, and thus require the minimum set of
dependencies. The runtime-compile package requires the kernels to be
compiled during runtime, and thus require the full set of opencl
tools, including clocl.

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>