thttpd: remove outdated package
thttpd has been outdated and not supported upstream for many years, which
could pose a major security risk. There is a supported fork of thttpd,
called sthttpd, and OE/Yocto provides a recipe in meta-webserver layer of
meta-openembedded, which can be added. Alternatively, lighttpd is a preferred
Web server provided by openembedded-core and is already used by Matrix.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>