packagegroup-arago-tisdk-addons: add ti-rpmsg-char for all K3 platforms
libti_rpmsg_char is a new utility library providing a simple API for
applications wanting to talk to remote processors over the rpmsg_char
Linux kernel driver with virtio-rpmsg back-end transport.

Add this library and examples for all the K3 platforms. It does not
support legacy SoCs at the moment, but this can be adjusted later on
when the support for the same is added to the library.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Gowtham Tammana <g-tammana@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>