statcol: New recipe for statistics collector application
statcol is a statistics collector application which can be used
to get various statistics data like CPU load, DDR bandwidth,
network speed, etc from various sources.

It allows to simply dump the data over command line or it can
also be used to send to a websocket for visualization on a
host machine.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Reviewed-by: Denys Dmytriyenko <denys@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>