branding-bleeding: optee-os: Stop auto rev of optee-os
Current mainline optee-os has the patches applied and the bleeding
build fails to apply the patches as they already exist on the current HEAD.
So optee-os should be pinned to 3.11 with the patches needed to build successfully.

Long term may need a recipe to build mainline HEAD exclusively and for
non-bleeding builds pin OPTEE-os to a specific version.

Signed-off-by: Dan Murphy <dmurphy@ti.com>