branding: Add branding-core-next to aid in migration
Add a core-next brand branding type to be able to re-direct the recipe
preferred providers or versions to highest PV'd recipe. The
branding-core should indicate the current version to build if not the
highest version.

For the purpose of an example the Arago distribution will be migrating
from the 5.4 kernel to the 5.10 during the 2021 time frame.  In order to
assist with a smoother migration the distribution needs to be able to
build the 5.4 kernel for the core brand and build the 5.10 kernel for
the core-next.

Once the transition from 5.4->5.10 kernel is done then the
PREFERRED_VERSION in the branding-core.inc can be removed.

From: Dan Murphy <dmurphy@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>