tisdk-tiny-image: remove packagedata task breaking initramfs packaging
The way initramfs gets included into other packages causes its packagedata
manifest to be absent in sstate. This used to be a warning and as part of
cleanup and proper fixing in master, became an error. That was inadvertently
backported to dunfell, but will be reverted soon. Regardless, fix it by
removing the unnecessary task in the initramfs image recipe.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>