tisdk-bundle: fix prebuilt-images U-Boot binary name for HS
HS builds do not generate a -${MACHINE} version of the u-boot.img binary.
Not sure why we deploy these named this way to begin with, maybe a left-
over from when we had multiple machines per SDK.. To fix the HS builds
lets ship the binary named as-is for all platforms.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>