oe-core: Fix rm_work issue with some packages from oe-core.
There appears to be an issue with a few packages from oe-core
and the rm_work feature when combined with the srcipk feature
of arago.  These recipes patch files and after the rm_work the
bitbake flow does not re-fetch the source again, so the patch
fails the second time around.  Easy fix is to just exclude
rm_worke on these packages.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>