meta-arago-extras: recipes-devtools: eventdump: Fix linking against local library
https://github.com/TI-ECS/eventdump unfortunately is a unmaintained
project.

LDFLAGS += -L$(NFSROOT)/lib should have dependent on NFSROOT variable.
Instead this forces the eventdump to break the sandbox and attempt to
link up against the local PC's lib - which is a NO-NO for many reasons.

Reported to have failed on a Fedora system.

Bump the PR up to indicate a fix up where we have proper linkage

Reported-by: Daniel Thompson <daniel.thompson@linaro.org>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>