meta-arago-distro: drop support for top-level Makefile builds in SDK
Historically, top-level TI-SDK Makefile was an extra layer on top of the
standard OE SDK to drive the cross-compilation of select components outside
of the normal OpenEmbedded/Yocto flow - back in the day when Yocto Project
was young and unfamiliar to embedded developers. Times have changed and the
OpenEmbedded/Yocto Project became the standard, while maintaining two build
frameworks in the same product requires a lot of effort, so let's drop the
top-level Makefile support going forward.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>