meta-arago-distro: drop support for handling source IPKs in SDK
As support for top-level Makefile in SDK is being dropped, there's no need to
generate source IPK packages to include sources in the SDK for cross-compiling.
Generation of source IPKs is a bit kludgy as it's not supported upstream or by
opkg, hence it often gets in the way of migrating the layer to the new release.
Therefore, remove sourceipk.bbclass, arago-source-ipk.inc package configs and
corresponding packagegroup-*-sdk-host.bb that pulled src packages to the SDK.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>