meta-arago: remove old deprecated long-SDKPATH hack
The long-SDKPATH hack was used for pre-allocating large enough fields
in the ELF headers of nativesdk binaries, so when they are relocated
the new path to the dynamic loader and libraries would fit. This is
no longer necessary for some time, so time to remove the hack.

Change the default SDKPATHINSTALL to include the SDK/distro version.
Also drop all workarounds that were needed due to the unusual SDKPATH.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>