qtwayland: Bump version to v5.15.7-lts-lgpl
There is a build error with the v5.15.4-lts-lgpl version and
xkbcommon. The bug was fixed in the next LTS, so go ahead and
move to the latest version available.

This haas been sent to meta-qt5 as a pull request.  When they
move to a version that does not have this build issue we will
revert this change.

Upstream-Status: Pending

Signed-off-by: Ryan Eatmon <reatmon@ti.com>