arago-prefs: Change how variants interact with the kernel choice
Rather than just explicitly locking the PREFERRED_PROVIDER_virtual/kernel
to linux-ti-staging and the suffix, use the :append to add the suffix to
whatever value PREFERRED_PROVIDER_virtual/kernel is.  This allows a
machine in a BSP layer to control which kernel to use as the bsse.

In the event that the kernel/variant choice is not valid the error
report from bitbake is very vague, but it does fail so you will not get
an incorrect build.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>