meta-toolchain-arago: fix internal toolchain installer install issue
When building the toolchain installer via 'bitbake -c populate_sdk ...'
the resulting toolchain installer throws an error complaining about it
can't find an 'environment-setup-*' file and hangs. The issue is caused
by how the TOOLCHAIN_HOST_TASK variable is populated by the Arago distro
layer, which must be extended via ':append', and not via '+='. See [1]
and [2] for more background on this.

[1] https://www.openembedded.org/pipermail/openembedded-core/2015-July/106931.html
[2] https://community.toradex.com/t/problem-with-angstrom-sdk-installer/5810/8

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>