ocl-gl-headers: Remove this recipe
OpenGL headers should be provided by the GPU driver. If that is not
available then they can be fetch from the official registry[0].

OpenCL does not build and needs reworked. When we restore it we should
use one of the above two methods instead of carrying the headers here.
Remove these headers and this recipe.

[0] https://registry.khronos.org/OpenGL/

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>