linux-libc-headers: Remove custom Linux libc headers
We do not need nor want a custom libc as the base for our packages.
Using our own Linux UAPI includes creates this situation and will
cause issues for packages that make use of the custom kernel headers
when moving from kernel to kernel.

If we do have some package that does need some custom header then
it should be listed in the package itself so this dependency can
be tracked correct. This is all described in the base recipe here[0].

As recommended, let's not create a custom linux-libc-headers. Remove
this recipe and references to specific versions.

[0] https://cgit.openembedded.org/openembedded-core/tree/meta/recipes-kernel/linux-libc-headers/linux-libc-headers.inc#n7

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>