glfw: convert to append and drop patches
A reasonable version of glfw is now provided by meta-oe. Also the EGL
version this package needs to build a against can be satisfied by Mesa
EGL so this old patch isn't required anymore.

Add a PACKAGECONFIG variable for wayland and use it. This append will
need to be adjusted again when meta-oe glfw patches settle.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>