arago.conf: bump distro version
There are some relatively major changes coming to meta-ti in the form
of kernel 6.1 LTS, 2023 U-boot with binman and reworking GP/HS machine
configs and firmware handling, as well as slight layer restructuring
in meta-arago itself. Arago distro version hasn't been bumped for a
while and it is a good time to do so now while things are considerably
stable.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>