arago.conf: drop the cairo gles override
Cairo OpenGLES selector logic has been moved into the bsp layer so that
other distros can get this optimization on our platforms as well.

Also it's necessary to drop this as not setting a gpudriver provider in
the meta-ti layer will create the exact clashing configs that extra
logic was created to resolve. Gotta love layer priority!

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>