opencl: stop using "virtual/" in RPROVIDES and RDEPENDS
Recipes shouldn't use the "virtual/" string in RPROVIDES and RDEPENDS.

That's confusing because "virtual/" has no special meaning in
RPROVIDES and RDEPENDS (unlike in PROVIDES and DEPENDS).

Instead, using "virtual-" instead of "virtual/"
as already done in the glibc recipe.

This is inspired by a bug report on oe-core [1], and implemented in a
patch to that repo [2].

[1] https://bugzilla.yoctoproject.org/show_bug.cgi?id=14538
[2] https://git.openembedded.org/openembedded-core/commit/meta/recipes-bsp/grub/grub-bootconf_1.00.bb?id=93ac180d8c389f16964bce8bd5538d9389e970e6

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>