weston-init: drop all the old append logic
First step in non-root weston.

Drop the following files:
	- weston-init/init
	- weston-init/runWeston
	- weston-init/ti43x/weston.ini
	- weston-init/wayland_env.sh

We previously had some append logic to remove the systemd-service
because people were only updating the sysVinit script, and the sysVinit
script was updated to include touch screen calibration logic. Now we can
drop this as weston now has it's own calibration routine per user.

In addition to this, this append file also introduced a runWeston
wrapper and a wayland_env script to setup environment variables for the
root user weston instance, all of which is now unnecessary as root
weston is generally not a good idea.

This append now only serves to offer a new weston.ini file to the
existing weston-init package with the hooks for libinput to create and
apply touchscreen calibrations for individual users.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>