arago-image.inc: fix the default systemd target
SYSTEMD_DEFAULT_TARGET is dynamically set based on the contents of
IMAGE_FEATURES. Would you like to know what doesn't get updated when we
use a bunch of package groups to dictate what is present in the image?

This makes systemd actually start graphical applications if the opengl
distro feature is set. This will need to be adjusted if further
graphical environments are added.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>