packagegroups: packagegroup-arago-tisdk-addons: Add sudo in arago filesystem
This commit adds sudo to the Arago filesystem. The addition of sudo is necessary due to weston running rootless.
Hence, Customers who are working on applications which requires weston may now need sudo for that extra privileges.
Furthermore, the inclusion of sudo is in response to past customer requests for this feature.

Signed-off-by: Chirag Shilwant <c-shilwant@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>