arago.conf: make things more readable and fix the virtuals
Make things more readable and reliable by using the built in boolean
check instead of python's type casting. Drop the array indexing in favor
of a more direct if else statement.

Also fix the virtual provider for login manager. There should only be 1
login provider and it should be shadow-base for systemd *or* busybox for
sysVinit systemd.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>