classes: tisdk-bundle: Move the licenses and manifests out of docs
We plan on not having docs populated in the intaller, to allow for
easier doc updates on the release pages without having to update the
installer. So move the licenses directory out of the docs directory.

Also, update manifest file location to a new directory for the same.
"docs/" directory will not be generated.

Signed-off-by: Aniket Limaye <a-limaye@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>