tisdk-bundle: Change name of image files when extracting rootfs files
There was a recent commit on oe-core master [1] that changes how the
names of the images files are built up.  This recipe uses a number of
variables to assembles the path to the images files so that it can
extract part of the rootfs.  We need to update that path to match
what upstream is now producing.

[1] https://git.openembedded.org/openembedded-core/commit/?id=26d97acc71379ab6702fa54a23b6542a3f51779c

Signed-off-by: Ryan Eatmon <reatmon@ti.com>