libcamera: Add recipe for v0.0.5
For kirkstone, an older recipe is present in meta-oe which is using
libcamera from December 2021 with many missing features.

TI's 6.1 based kernel can support the simple pipeline of libcamera, and
benefit from the latest features in release v0.0.5, so we override the
recipe here for arago.

The new recipe here is taken as-is from meta-oe master.

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>